rename Connection to CalendarConnection
